Create demoCodeCopier.py
Initial check in, pretty stubbed out at this point. But on a linux box, this can be used to copy some block of code into the middle click clipboard, based on the numeric input. Like if you're entering steps in a demo, this will copy the first block of code, second block of code, etc for the demo into the clipboard. Hopefully this will be useful for streamlining walkthroughs and demos of building up code over time.